api/requirements.txt
# FastAPI & Web
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9

# Database
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.0

# Validation
pydantic==2.9.0
pydantic-settings==2.5.0
email-validator==2.1.0  # For EmailStr

# Authentication
passlib[bcrypt]==1.7.4  # Password hashing
python-jose[cryptography]==3.3.0  # JWT tokens
bcrypt==4.1.2  # Bcrypt backend

# HTTP Client
httpx==0.27.0

# Caching & Rate Limiting
redis==5.0.0
slowapi==0.1.9

# Document Generation
docxtpl==0.18.0
python-docx>=1.1.1

# Encryption (for sensitive data)
cryptography==42.0.0
